+++ sketch_feb18a/spider_sense.h
@@ -19,7 +19,8 @@ extern int defineLevel(int distance);   // the function take in input the distan
 int pwmButton(int level);               // the function return an impulse for the buttons based on the level
 int pwmBeeper(int level);               // the function return an impulse for the beeper based on the level
 
-double avarage2(double media_precedente, int n, int nuovo_valore);
+double avarage2(double media_precedente, int n, double nuovo_valore);
 double deriveHeight(double *arr, int size);
+
 
 #endif
